WIP: scripts/maintenance: Per-resource AWS cleanup scripts
I'm not familiar enough with grafiti to want to run the old
clean-aws.sh in production.  This commit lands a collection of
per-resource cleaners, as well as an 'all' script that ties them all
together in what seems like a reasonable order.  I'm generally trying
to use the resource creation date with a four-hour expiration to
decide if a resource is stale.  And for resources that don't expose a
creation date, I'm assuming they're stale if they're still around
after an hour (our Prow tasks are running in their own CI-only AWS
account, so we don't have to worry about long-running resources).  I'm
also using jq for filtering, because that's what I was familiar with.
It would be better to use JMESPath and --query to perform these
filters.

It will me much more reliable (and faster) if we can migrate these to
use the new expirationDate tag from openshift/release@f64b8eaf
(cluster-launch-installer-e2e: Set expirationDate tags, 2018-07-26,
openshift/release#1103), but I haven't gotten around to that yet.  I'd
also like to use something like:

  $ aws resourcegroupstaggingapi get-resources --query "ResourceTagMappingList[?Tags[? Key == 'expirationDate' && Value < '$(date --utc -d '4 hours' '+%Y-%m-%dT%H:%M')']].ResourceARN" --output text

to get all expired resources in one go regardless of type, but haven't
had time to get that settled either.