Define a map of region to instance "classes" for regions that don't support m4
The only instance type available in all AWS regions today is i3. We prefer m4
because of the higher EBS attach limit. But for regions that don't support m4
we use m5.

This also adds a CI job which checks the AWS pricing API to keep our region
map up 2 date.