pkg/types/aws/defaults/platform: Per-region instance-class defaults
From Trevor:

For regions like eu-west-3 (Paris) that don't support m4.  The only
instance type available in all AWS regions today is i3.  We prefer m4
because of the higher EBS attach limit.  But for regions that don't
support m4 we use m5.

This also adds a CI job which checks the AWS pricing API to keep our
region map up to date.

The eu-north-1 region is new as of 2018-12-12 [1], and it show up in:

  $ aws ec2 describe-regions --query 'Regions[].RegionName' --output json | jq -r 'sort[]'
  ...
  eu-central-1
  eu-north-1
  eu-west-1
  ...

The GovCloud entries didn't show up in that query (maybe they are only
listed for accounts that have access?), but they are listed on [2].

The new test is in platformtests/aws to keep it out of our default
unit tests.  Maintainers can run it manually to verify that the
default fallbacks we hard-code are still appropriate for the current
AWS offerings.  It's not in tests/aws, because 'dep ensure' seems to
ignore tests.  I think it's fine to have both platformtests and tests
in parallel, with the former holding tests who share the project
vendor/ and the latter holding tests which have a distinct vendor/.
bdd-smoke is in tests with its own vendor/ intentionally to keep
gomega and other test-specific libraries out of the project vendor/
[3].  But for these tests we're using Go's built-in test framework, so
it's less maintenance to layer our minimal additional dependencies
into the project vendor/.

[1]: https://aws.amazon.com/blogs/aws/now-open-aws-europe-stockholm-region/
[2]: https://aws.amazon.com/about-aws/global-infrastructure/regional-product-services/
[3]: https://github.com/openshift/installer/pull/552#issuecomment-459892007