release: Allow release image to be directly substituted into binary
As part of our release process we build container images for installer
that are added to the release image (which has a cryptographic
relationship to the images it contains, giving strong integrity). A
consumer should be able to download and install a locked installer
binary that uses the payload. However, we would prefer to not
rebuild the binary outside of the image, but instead have:

1. a source for the binary from the payload
2. the binary be locked to the payload it comes from

This commit allows a build system above the payload to extract the
installer binary for linux from the image (other platforms later)
and perform a replacement on the binary itself, patching:

```
_RELEASE_IMAGE_LOCATION_XXXXXXXXX...
```

with

```
quay.io/openshift/ocp-release:v4.0\x00
```

without requiring a recompilation of the binary. The internal code
checks the constant and verifies bounds (panicking if necessary)
and then returns the updated constant. This allows a simpler
replacement process to customize a binary for both external use
and offline use that is locked to a payload.