pkg/destroy/aws/aws: Promote "Terminated" to info-level log
We spend a while after the "Terminating" log waiting for the instances
to go down.  By promoting this to an info-level log, folks running at
that level can see the instances gradually completing termination and
have a more obvious feeling of progress (instead of feeling like the
destroy code may be hung up ;).